Debugging for updated packages and numerical issues. 1. Tested with python 3.11.5 and lmfit 1.2.2 2. Had issues with nans. We keep nans in data during line fitting to have continuous points. Have to set nan_policy='omit' to ignore these points. But other nans occurring due to numerical issues, which  can see more explicitly by setting nan_policy='raise' (default). These were occurring in continuum and line fits in various contexts. To deal with this, changed lower bounds wherever possible to machine epsilon (~2.2e-16 for double-precision float64) rather than 0. Seems to have worked. 3. Changed back to explictly casting everything as float64, which lmfit seems to want (and makes more sense anyway). The tolerances seem to function more as expected now. 4. Added lmfit argument options to questfit. 5. More explictly commented questfit and made a few cosmetic code changese. Moved readcf to questfitfcn.py. 6. A few minor bugfixes.

+++ q3dfit/fitloop.py
@@ -156,7 +156,7 @@ def fitloop(ispax, colarr, rowarr, cube, q3di, listlines, specConv,
                         listlinesz[line] = \
                             np.array(listlines['lines']
                                      [(listlines['name'] == line)],
-                                     dtype='float32') * \
+                                     dtype='float64') * \
                                 (1. + q3di.zinit_gas[line][i, j, ])
 
             if q3di.docontfit:
